Refactor shared schema parser helpers
